Clean up NcWMSColorbarControl
